+++ src/lib/icons/Info.svelte
@@ -1,6 +1,6 @@
 <svg
 	xmlns="http://www.w3.org/2000/svg"
-	class="fill-cyan-600"
+	class="fill-cyan-500"
 	height="20"
 	width="20"
 	viewBox="0 0 512 512"